# backend/docs/openapi.json
{
    "openapi":  "3.0.3",
    "info":  {
                 "title":  "Portfolio 3.0 API",
                 "description":  "CHRONO//ROOTS backend API serving verified portfolio content.",
                 "version":  "1.0.0"
             },
...
    "paths":  {
...
                  "/api/v1/projects/{slug}":  {
                                                  "get":  {
                                                              "parameters":  [
                                                                                 {
                                                                                     "schema":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                     "in":  "path",
                                                                                     "name":  "slug",
                                                                                     "required":  true
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Default Response"
                                                                                        }
                                                                            }
                                                          }
                                              },
...
                  "/api/v1/creative/{slug}":  {
                                                  "get":  {
                                                              "parameters":  [
                                                                                 {
                                                                                     "schema":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                     "in":  "path",
                                                                                     "name":  "slug",
                                                                                     "required":  true
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Default Response"
                                                                                        }
                                                                            }
                                                          }
                                              },
...
              },
    "servers":  [
                    {
                        "url":  "http://localhost:3001",
                        "description":  "Local development"
                    }
                ]
}